+++ src/stars/EAGLE/stars_part.h
@@ -76,14 +76,14 @@ struct spart {
   union {
 
     /*! Birth time */
-    float birth_time;
+    double birth_time;
 
     /*! Birth scale factor */
-    float birth_scale_factor;
+    double birth_scale_factor;
   };
 
   /*! Scale-factor / time at which this particle last did enrichment */
-  float last_enrichment_time;
+  double last_enrichment_time;
 
   /*! Initial star mass */
   float mass_init;
